Split sku types
