+++ packages/types/src/index.ts
@@ -14,6 +14,7 @@ export * from './discord/invites.js'
 export * from './discord/messages.js'
 export * from './discord/polls.js'
 export * from './discord/reference.js'
+export * from './discord/skus.js'
 export * from './discord/webhookEvents.js'
 export * from './discordeno.js'
 export * from './shared.js'
